Move goal selector bridge out of mixin namespace
